+++ services/cloudfront/src/main/java/software/amazon/awssdk/services/cloudfront/internal/utils/SigningUtils.java
@@ -17,6 +17,9 @@ package software.amazon.awssdk.services.cloudfront.internal.utils;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
+import java.io.InputStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
 import java.security.PrivateKey;
@@ -26,9 +29,14 @@ import java.security.SignatureException;
 import java.time.Instant;
 import java.util.Base64;
 import software.amazon.awssdk.annotations.SdkInternalApi;
+import software.amazon.awssdk.core.exception.SdkClientException;
+import software.amazon.awssdk.services.cloudfront.internal.auth.Pem;
+import software.amazon.awssdk.services.cloudfront.internal.auth.Rsa;
+import software.amazon.awssdk.utils.IoUtils;
+import software.amazon.awssdk.utils.StringUtils;
 
 @SdkInternalApi
-public class SigningUtils {
+public final class SigningUtils {
 
     private SigningUtils() {
     }
@@ -128,6 +136,70 @@ public class SigningUtils {
         } catch (NoSuchAlgorithmException | SignatureException e) {
             throw new IllegalStateException(e);
         }
+    }
+
+    /**
+     * Generate a policy document that describes custom access permissions to
+     * apply via a private distribution's signed URL.
+     *
+     * @param resourceUrl
+     *            The HTTP/S resource path that restricts which distribution and
+     *            S3 objects will be accessible in a signed URL, i.e.,
+     *            <tt>"https://" + distributionName + "/" + objectKey</tt> (may
+     *            also include URL parameters). The '*' and '?' characters can
+     *            be used as a wildcards to allow multi-character or single-character
+     *            matches respectively:
+     *            <ul>
+     *            <li><tt>*</tt> : All distributions/objects will be accessible</li>
+     *            <li><tt>a1b2c3d4e5f6g7.cloudfront.net/*</tt> : All objects
+     *            within the distribution a1b2c3d4e5f6g7 will be accessible</li>
+     *            <li><tt>a1b2c3d4e5f6g7.cloudfront.net/path/to/object.txt</tt>
+     *            : Only the S3 object named <tt>path/to/object.txt</tt> in the
+     *            distribution a1b2c3d4e5f6g7 will be accessible.</li>
+     *            </ul>
+     * @param activeDate
+     *            An optional UTC time and date when the signed URL will become
+     *            active. If null, the signed URL will be active as soon as it
+     *            is created.
+     * @param expirationDate
+     *            The UTC time and date when the signed URL will expire. REQUIRED.
+     * @param limitToIpAddressCidr
+     *            An optional range of client IP addresses that will be allowed
+     *            to access the distribution, specified as an IPv4 CIDR range
+     *            (IPv6 format is not supported). If null, the CIDR will be omitted
+     *            and any client will be permitted.
+     * @return A policy document describing the access permission to apply when
+     *         generating a signed URL.
+     */
+    public static String buildCustomPolicyForSignedUrl(String resourceUrl,
+                                                       Instant activeDate,
+                                                       Instant expirationDate,
+                                                       String limitToIpAddressCidr) {
+        if (expirationDate == null) {
+            throw SdkClientException.create("Expiration date must be provided to sign CloudFront URLs");
+        }
+        if (resourceUrl == null) {
+            resourceUrl = "*";
+        }
+        return buildCustomPolicy(resourceUrl, activeDate, expirationDate, limitToIpAddressCidr);
+    }
+
+    /**
+     * Creates a private key from the file given, either in pem or der format.
+     * Other formats will cause an exception to be thrown.
+     */
+    public static PrivateKey loadPrivateKey(Path keyFile) throws Exception {
+        if (StringUtils.lowerCase(keyFile.toString()).endsWith(".pem")) {
+            try (InputStream is = Files.newInputStream(keyFile)) {
+                return Pem.readPrivateKey(is);
+            }
+        }
+        if (StringUtils.lowerCase(keyFile.toString()).endsWith(".der")) {
+            try (InputStream is = Files.newInputStream(keyFile)) {
+                return Rsa.privateKeyFromPkcs8(IoUtils.toByteArray(is));
+            }
+        }
+        throw SdkClientException.create("Unsupported file type for private key");
     }
 
 }
