+++ services/cloudfront/src/main/resources/codegen-resources/customization.config
@@ -6,5 +6,10 @@
         "listFieldLevelEncryptionProfiles",
         "listPublicKeys",
         "listStreamingDistributions"
-    ]
+    ],
+    "utilitiesMethod": {
+        "returnType": "software.amazon.awssdk.services.cloudfront.CloudFrontUtilities",
+        "createMethodParams": [
+        ]
+      }
 }
